backend/requirements.txt
fastapi
uvicorn
yt-dlp
google-generativeai
python-multipart
python-dotenv
